% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{AUDUSD2015_H1}
\alias{AUDUSD2015_H1}
\title{Hourly AUD/USD example data}
\format{
A data frame with 6226 rows and 7 columns: Date, Timestamp, Open,
  High, Low, Close, and Volume.
}
\usage{
data(AUDUSD2015_H1)
}
\description{
AUD/USD observations supplied with CopulaOne.
}
\keyword{datasets}
