% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finTools.R
\name{autoARMAGARCH}
\alias{autoARMAGARCH}
\title{Auto select ARMA+GARCH}
\usage{
autoARMAGARCH(dat, distvec = c("norm", "snorm", "std", "sstd", "ged", "sged",
  "nig", "ghyp", "jsu"), maxAR = 4, maxMA = 4, pqGARCH = c(1, 1),
  windowLength = 100, sinkflag = T)
}
\arguments{
\item{dat:}{input of univariate time series data.}

\item{distvec:}{vector of names of distributions to choose from for innovations}

\item{maxAR:}{maximum AR order}

\item{maxMA:}{maximum MA order}

\item{pqGARCH:}{specify GARCH orders (default: c(1,1))}

\item{windowLength:}{rolling window length}

\item{sinkflag:}{whether sink outputs (default: T)}
}
\description{
automatically delect orders for ARMA, and distributions for innovations of GARCH
}
\examples{
data("USDCAD2015_H1")
dat <- USDCAD2015_H1$Close[1:250] 
logr <- diff(log(dat))
autoARMAGARCH(dat=logr, windowLength=245)
}
\keyword{ARMA}
\keyword{GARCH}
\keyword{auto}

