% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{betaFRA1_COP}
\alias{betaFRA1_COP}
\title{Blomqvist's beta of the FRA1 copula}
\usage{
betaFRA1_COP(eta, theta)
}
\arguments{
\item{eta}{Lower-tail parameter in [-1,1).}

\item{theta}{Upper-tail parameter in [-1,1).}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Blomqvist's beta of the FRA1 copula
}
\examples{
betaFRA1_COP(eta = -0.2, theta = 0.3)

}
\keyword{distribution}
