% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CopulaOne-package}
\alias{CopulaOne}
\alias{CopulaOne-package}
\title{Full-range tail dependence copulas}
\description{
Distribution functions, simulation, dependence measures, and likelihood
fitting for the FRA1, GGEE, and PPPP bivariate copula families.
}
\author{
\strong{Maintainer}: Lei Hua \email{larry.lei.hua@gmail.com}

Authors:
\itemize{
  \item Lei Hua \email{larry.lei.hua@gmail.com}
}

}
\keyword{internal}
