% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dGGEE}
\alias{dGGEE}
\title{Density of univariate margins of the GGEE model}
\usage{
dGGEE(x, al, be, maxit = 1e+05)
}
\arguments{
\item{x}{Numeric marginal observations.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{maxit}{Positive maximum number of numerical iterations. Nonconvergence is reported.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Density of univariate margins of the GGEE model
}
\examples{
dGGEE(3, 1, 2)
}
\keyword{distribution}
