% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dPP}
\alias{dPP}
\title{density of univariate margins of the PP model (i.e., Pareto I / Pareto I)}
\usage{
dPP(x, ga1, ga2)
}
\arguments{
\item{x}{Numeric marginal observations.}

\item{ga1}{Positive lower-tail shape parameter.}

\item{ga2}{Positive upper-tail shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
density of univariate margins of the PP model (i.e., Pareto I / Pareto I)
}
\examples{
plot(sapply(seq(0,10, length=50), function(x){dPP(x, 2,2)}), type="l")
}
\keyword{distribution}
