% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dPPPP_COP}
\alias{dPPPP_COP}
\title{Copula Density Function - PPPP_COP}
\usage{
dPPPP_COP(u, v, al, be, a, b)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{v}{Second copula argument.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
PPPP_COP - Copula density function of the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
dPPPP_COP(0.2, 0.4, 1,1,2,2)
}
\keyword{distribution}
