% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dPPPP_COP_1_90}
\alias{dPPPP_COP_1_90}
\title{Copula Density Function of PPPP_COP that is rotated for 90 degrees counter clock-wise (when a=b=1)}
\usage{
dPPPP_COP_1_90(uvec, vvec, al, be)
}
\arguments{
\item{uvec}{First copula arguments.}

\item{vvec}{Second copula arguments.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Copula density function of PPPP_COP that is rotated for 90 degrees counter clock-wise (when a=b=1)
}
\examples{
dPPPP_COP_1_90(0.2, 0.4, 1,1)
dPPPP_COP_1_90(0.8, 0.3, 1,1)
dPPPP_COP_1_90(c(0.2,0.8), c(0.4, 0.3), 1,1)
}
\keyword{distribution}
