% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jdPPPP}
\alias{jdPPPP}
\title{Joint density of the PPPP model}
\usage{
jdPPPP(x1, x2, al, be, a, b)
}
\arguments{
\item{x1}{First marginal observation.}

\item{x2}{Second marginal observation.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Joint density of the PPPP model
}
\examples{
jdPPPP(10,2, 1, 1,2,2)
}
\keyword{distribution}
