% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jpPPPP}
\alias{jpPPPP}
\title{Joint CDF of the PPPP model}
\usage{
jpPPPP(x1, x2, al, be, a, b)
}
\arguments{
\item{x1}{First marginal observation.}

\item{x2}{Second marginal observation.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Joint CDF of the PPPP model
}
\examples{
jpPPPP(2,2, 1.2, 0.8, 1, 1)
}
\keyword{distribution}
