% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPPPP}
\alias{pPPPP}
\title{CDF of univariate margins of the PPPP model}
\usage{
pPPPP(x, al, be, a, b, log = FALSE)
}
\arguments{
\item{x}{Numeric marginal observations.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}

\item{log}{Logical; return the logarithm of the CDF.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
CDF of univariate margins of the PPPP model
}
\examples{
plot(sapply(seq(0.0001,20, length=50), function(x){pPPPP(x, 1,1,2,2,log=FALSE)}), type="l",ylab="")
}
\keyword{distribution}
