% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPPPP_COP}
\alias{pPPPP_COP}
\title{Joint CDF of the PPPP copula model}
\usage{
pPPPP_COP(u, v, al, be, a, b)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{v}{Second copula argument.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Joint CDF of the PPPP copula model
}
\examples{
pPPPP_COP(0.9, 0.3, 1.2, 0.8, 1, 1)
}
\keyword{distribution}
