% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPPPP_COP_1}
\alias{pPPPP_COP_1}
\title{Joint CDF of the PPPP copula model (when a=b=1)}
\usage{
pPPPP_COP_1(u, v, al, be)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{v}{Second copula argument.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Joint CDF of the PPPP copula model (when a=b=1)
}
\examples{
pPPPP_COP_1(0.9, 0.3, 1.2, 0.8)
}
\keyword{distribution}
