% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperEV.R
\name{pUEV_GGEE_COP}
\alias{pUEV_GGEE_COP}
\title{Joint cdf of the upper extreme value copula of the GGEE copula}
\usage{
pUEV_GGEE_COP(u, v, b)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{v}{Second copula argument.}

\item{b}{Dependence parameter; 0 < b < 1 uses the extreme-value formula, otherwise the CDF is u*v.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Joint cdf of the upper extreme value copula of the GGEE copula
}
\examples{
pUEV_GGEE_COP(0.3, 0.4, 1.2)
}
\keyword{distribution}
