% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{qGGEE}
\alias{qGGEE}
\title{Quantile of univariate margins of the GGEE model}
\usage{
qGGEE(u, al, be, maxit = 1e+05)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{maxit}{Positive maximum number of numerical iterations. Nonconvergence is reported.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Quantiel of univariate margins of the GGEE model
}
\examples{
qGGEE(0.9, 1, 1)
}
\keyword{distribution}
