% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{qPP}
\alias{qPP}
\title{quantile of univariate margins of the PP model (i.e., Pareto I / Pareto I)}
\usage{
qPP(u, ga1, ga2)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{ga1}{Positive lower-tail shape parameter.}

\item{ga2}{Positive upper-tail shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
quantile of univariate margins of the PP model (i.e., Pareto I / Pareto I)
}
\examples{
plot(sapply(seq(0.99,1-0.001, length=1000), function(x){qPP(x, 0.2,0.3)}), type="l")
}
\keyword{distribution}
