% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{qPPPP}
\alias{qPPPP}
\title{quantile of univariate margins of the PPPP model}
\usage{
qPPPP(u, al, be, a, b, maxit = 1e+05)
}
\arguments{
\item{u}{Probability or first copula argument.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}

\item{maxit}{Positive maximum number of numerical iterations. Nonconvergence is reported.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
quantile of univariate margins of the PPPP model
}
\examples{
\dontrun{
plot(sapply(seq(0.001,0.999, length=100), function(x){qPPPP(x, 0.3, 1.3, 1, 1)}), type="l",ylab="")
}
}
\keyword{distribution}
