% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{rGGEE_COP}
\alias{rGGEE_COP}
\title{Sampling based on the full-range tail dependence copula}
\usage{
rGGEE_COP(n, al, be, seed = NULL, maxit = 1e+05)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{seed}{Optional integer random seed.}

\item{maxit}{Positive maximum number of numerical iterations. Nonconvergence is reported.}
}
\value{
A two-column numeric matrix of uniform scores.
}
\description{
Generating random samples based on the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
rGGEE_COP(20, 1.2, 0.2, seed = 100)
}
\keyword{distribution}
