% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{rPPPP_COP}
\alias{rPPPP_COP}
\title{Sampling based on the PPPP copula}
\usage{
rPPPP_COP(n, al, be, a, b, seed = NULL)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}

\item{seed}{Optional integer random seed.}
}
\value{
A two-column numeric matrix of uniform scores.
}
\description{
Generating random samples based on the PPPP copula
}
\examples{
plot(rPPPP_COP(100, 1.2, 0.2, 1, 1, seed = 100))
}
\keyword{distribution}
