% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tailFRA1_COP}
\alias{tailFRA1_COP}
\title{Tail coefficients and orders of the FRA1 copula}
\usage{
tailFRA1_COP(eta, theta)
}
\arguments{
\item{eta}{Lower-tail parameter in [-1,1).}

\item{theta}{Upper-tail parameter in [-1,1).}
}
\value{
A named numeric vector of dependence summaries.
}
\description{
Tail coefficients and orders of the FRA1 copula
}
\keyword{distribution}
