% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauFRA1_COP}
\alias{tauFRA1_COP}
\title{Kendall's tau of the FRA1 copula}
\usage{
tauFRA1_COP(eta, theta, quad_n = 64)
}
\arguments{
\item{eta}{Lower-tail parameter in [-1,1).}

\item{theta}{Upper-tail parameter in [-1,1).}

\item{quad_n}{Number of Gauss-Legendre nodes, an integer of at least 2.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Kendall's tau of the FRA1 copula
}
\examples{
tauFRA1_COP(eta = -0.2, theta = 0.3)

}
\keyword{distribution}
