% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauGGEE_COP}
\alias{tauGGEE_COP}
\title{Kendall's tau of the GGEE copula}
\usage{
tauGGEE_COP(al, be)
}
\arguments{
\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Kendall's tau of the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
tauGGEE_COP(0.5, 1)
}
\keyword{distribution}
