% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauGGEE_COP_sim}
\alias{tauGGEE_COP_sim}
\title{Kendall's tau of the GGEE copula (based on simulations)}
\usage{
tauGGEE_COP_sim(al, be, n = 1e+05)
}
\arguments{
\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{n}{Number of observations to generate.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Kendall's tau of the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
tauGGEE_COP_sim(0.5, 1, n=10000)
}
\keyword{distribution}
