% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauPPPP_COP}
\alias{tauPPPP_COP}
\title{Kendall's tau of the PPPP copula}
\usage{
tauPPPP_COP(al, be, a, b)
}
\arguments{
\item{al}{Positive first shape parameter.}

\item{be}{Positive second shape parameter.}

\item{a}{Positive shape parameter.}

\item{b}{Positive shape parameter.}
}
\value{
Numeric function values. Invalid parameters or numerical failures are reported.
}
\description{
Kendall's tau of the PPPP copula
}
\examples{
\dontrun{
tauPPPP_COP(3, 6, 5.9, 6.7)
}
}
\keyword{distribution}
