% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{USDCAD2015_H1}
\alias{USDCAD2015_H1}
\title{Hourly USD/CAD example data}
\format{
A data frame with 6226 rows and 7 columns: Date, Timestamp, Open,
  High, Low, Close, and Volume.
}
\usage{
data(USDCAD2015_H1)
}
\description{
USD/CAD observations supplied with CopulaOne.
}
\keyword{datasets}
