% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{C2GGEE_COP}
\alias{C2GGEE_COP}
\title{Conditional GGEE copula}
\usage{
C2GGEE_COP(u, v, al, be, maxit = 1e+05)
}
\arguments{
\item{u, v:}{data input.}

\item{al, be:}{parameters.}
}
\description{
Partial derivative wrt the second argument of the GGEE copula
}
\examples{
C2GGEE_COP(0.2, 0.6, al = 1.2, be = 0.8)
}
\keyword{conditional}
\keyword{copula}
