% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dGGEE}
\alias{dGGEE}
\title{Density of univariate margins of the GGEE model}
\usage{
dGGEE(x, a, b)
}
\arguments{
\item{a, b:}{parameters.}

\item{x:}{data input.}
}
\description{
Density of univariate margins of the GGEE model
}
\examples{
dGGEE(3, 1, 2)
}
\keyword{Density}

