% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dPPPP}
\alias{dPPPP}
\title{Density of univariate margins of the PPPP model (a!=al, b!=be)}
\usage{
dPPPP(x, al, be, a, b)
}
\arguments{
\item{al, be, a, b:}{parameters.}

\item{x:}{data input.}
}
\description{
Density of univariate margins of the PPPP model (a!=al, b!=be)
}
\examples{
plot(sapply(seq(0,20, length=100), function(x){dPPPP(x, 1,1,2,2)}), type="l",ylab="")
}
\keyword{Density}
