% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dPPPP_COP}
\alias{dPPPP_COP}
\title{Copula Density Function - PPPP_COP}
\usage{
dPPPP_COP(u, v, al, be, a, b)
}
\arguments{
\item{u, v}{values in (0,1).}

\item{al, be, a, b}{the four shape parameters.}
}
\description{
PPPP_COP - Copula density function of the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
dPPPP_COP(0.2, 0.4, 1,1,2,2)
}
\keyword{copula}
\keyword{density}
