% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{dPPPP_COP_1_90}
\alias{dPPPP_COP_1_90}
\title{Copula Density Function of PPPP_COP that is rotated for 90 degrees counter clock-wise (when a=b=1)}
\usage{
dPPPP_COP_1_90(uvec, vvec, al, be)
}
\arguments{
\item{al, be}{the four shape parameters.}

\item{u, v}{values in (0,1).}
}
\description{
Copula density function of PPPP_COP that is rotated for 90 degrees counter clock-wise (when a=b=1)
}
\examples{
dPPPP_COP_1_90(0.2, 0.4, 1,1)
dPPPP_COP_1_90(0.8, 0.3, 1,1)
dPPPP_COP_1_90(c(0.2,0.8), c(0.4, 0.3), 1,1)
}
