% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperEV.R
\name{dUEV_GGEE_COP}
\alias{dUEV_GGEE_COP}
\title{Density function of the upper extreme value copula}
\usage{
dUEV_GGEE_COP(u, v, b)
}
\arguments{
\item{u, }{v: inputs for the extreme value copula}

\item{b:}{the parameter of the extreme value copula}
}
\description{
Density of the upper extreme value copula of the full-range tail dependence copula
}
\examples{
dUEV_GGEE_COP(0.3, 0.4, 1.2)
}
\keyword{copula}
\keyword{extreme}
\keyword{upper}
\keyword{value}

