% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperEV.R
\name{dUEV_PPPP_COP}
\alias{dUEV_PPPP_COP}
\title{Density function of the upper extreme value copula for the PPPP copula}
\usage{
dUEV_PPPP_COP(u1, u2, be, a, b)
}
\arguments{
\item{be, a, b:}{the parameter of the extreme value copula}

\item{u, }{v: inputs for the extreme value copula}
}
\description{
Density function of the upper extreme value copula for the PPPP copula
}
\examples{
dUEV_PPPP_COP(0.3, 0.4, 0.6, 1, 1)
}
\keyword{copula}
\keyword{extreme}
\keyword{upper}
\keyword{value}
