% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaone.R
\name{fitCopulaOne}
\alias{fitCopulaOne}
\title{Model fitting with CopulaOne}
\usage{
fitCopulaOne(par, dat, flag = 1, opt = "L-BFGS-B", se = F,
  lower = c(0.1, 0.1), upper = c(5, 5), trace = 0, factr = 1e+09,
  printlevel = 0)
}
\arguments{
\item{trace, }{printlevel: integers for optim() and nlm(), respectively. (default: 0,0)}

\item{par:}{initial parameters for (a, b)}

\item{dat:}{input of data.}

\item{flag:}{indicate which numerical method for the appell function (default: flag = 1)}

\item{se:}{whether standard errors of parameters are reported (default: se = F)}
}
\description{
Fit bivariate data with the full-range tail dependence copula based on maximum likelihood
}
\examples{
# DO NOT run, and it takes time!
data("euro0306")
dat <- uscore(euro0306[,c(2,3)])[1:100,]
par <- c(0.3, 0.3)
fit <- fitCopulaOne(par, dat)
}
\keyword{fitting}

