% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jdGGEE}
\alias{jdGGEE}
\title{Joint density of the GGEE model}
\usage{
jdGGEE(x1, x2, al, be, flag = 1, integration = F)
}
\arguments{
\item{x1, x2:}{data input.}

\item{al, be:}{parameters.}

\item{integration}{if T: use numerical integration instead of appellf1, default is F}

\item{flag:}{flag used in appell::appellf1()}
}
\description{
Joint density of the GGEE model
}
\examples{
jdGGEE(10,2, 1, 1)
jdGGEE(10,2, 1, 1, integration = TRUE)
}
\keyword{Joint}
\keyword{density}
