% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jdGGEE}
\alias{jdGGEE}
\title{Joint density of the GGEE model}
\usage{
jdGGEE(x1, x2, a, b, flag = 1)
}
\arguments{
\item{x1, x2:}{data input.}

\item{a, b:}{parameters.}

\item{flag:}{flag used in appell::appellf1()}
}
\description{
Joint density of the GGEE model
}
\examples{
jdGGEE(10,2, 1, 1)
}
\keyword{Joint}
\keyword{density}

