% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jdPPPP}
\alias{jdPPPP}
\title{Joint density of the PPPP model}
\usage{
jdPPPP(x1, x2, al, be, a, b)
}
\arguments{
\item{x1, x2:}{data input.}

\item{al, be, a, b:}{parameters.}
}
\description{
Joint density of the PPPP model
}
\examples{
jdPPPP(10,2, 1, 1,2,2)
}
\keyword{Joint}
\keyword{density}
