% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jpGGEE}
\alias{jpGGEE}
\title{Joint CDF of the GGEE model}
\usage{
jpGGEE(x1, x2, a, b, flag = 1)
}
\arguments{
\item{x1, x2:}{data input.}

\item{a, b:}{parameters.}
}
\description{
Joint CDF of the GGEE model
}
\examples{
jpGGEE(0.2, 0.4, 1, 1)
}
\keyword{CDF}
\keyword{Joint}

