% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jpGGEE}
\alias{jpGGEE}
\title{Joint CDF of the GGEE model}
\usage{
jpGGEE(x1, x2, al, be, flag = 1, integration = F, maxit = 1e+05)
}
\arguments{
\item{x1, x2:}{data input.}

\item{al, be:}{parameters.}
}
\description{
Joint CDF of the GGEE model
}
\examples{
jpGGEE(0.2, 0.4, 1, 1)
jpGGEE(0.2, 0.4, 1, 1, integration = TRUE)
}
\keyword{CDF}
\keyword{Joint}
