% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{jpPPPP}
\alias{jpPPPP}
\title{Joint CDF of the PPPP model}
\usage{
jpPPPP(x1, x2, al, be, a, b)
}
\arguments{
\item{x1, x2:}{data input.}

\item{al, be, a, b:}{parameters.}
}
\description{
Joint CDF of the PPPP model
}
\examples{
jpPPPP(2,2, 1.2, 0.8, 1, 1)
}
\keyword{CDF}
\keyword{Joint}
