% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pGGEE}
\alias{pGGEE}
\title{CDF of univariate margins of the GGEE model}
\usage{
pGGEE(x, al, be, maxit = 1e+05)
}
\arguments{
\item{al, be:}{parameters.}

\item{x:}{data input.}
}
\description{
CDF of univariate margins of the GGEE model
}
\examples{
pGGEE(2, 1, 2)
}
\keyword{CDF}
