% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaone.R
\name{plotCopulaOne}
\alias{plotCopulaOne}
\title{Contour plots of CopulaOne}
\usage{
plotCopulaOne(a, b, marg = "normal", flag = 1, resolution = 30)
}
\arguments{
\item{a, }{b: parameters}

\item{marg:}{indicate normal scores or uniform scores (default: marg = "normal")}

\item{resolution:}{number of evaluations for each dimension (default is 30)}

\item{flag:}{used in the appell function (default is 1)}
}
\description{
Contour plots based on either normal scores or uniform scores
}
\examples{
plotCopulaOne(a=0.5, b=1.8)
plotCopulaOne(a=0.5, b=1.8, marg="uniform", resolution=20)
}
\keyword{contour}
\keyword{plot}

