% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPP}
\alias{pPP}
\title{CDF of univariate margins of the PP model (i.e., Pareto I / Pareto I)}
\usage{
pPP(x, ga1, ga2)
}
\arguments{
\item{x:}{data input.}

\item{al, be:}{parameters.}
}
\description{
CDF of univariate margins of the PP model (i.e., Pareto I / Pareto I)
}
\examples{
plot(sapply(seq(0,10, length=50), function(x){pPP(x, 2,2)}), type="l")
}
\keyword{CDF}
