% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPPPP}
\alias{pPPPP}
\title{CDF of univariate margins of the PPPP model}
\usage{
pPPPP(x, al, be, a, b, log = F)
}
\arguments{
\item{al, be, a, b:}{parameters.}

\item{x:}{data input.}

\item{log:}{return log or not}
}
\description{
CDF of univariate margins of the PPPP model
}
\examples{
plot(sapply(seq(0.0001,20, length=50), function(x){pPPPP(x, 1,1,2,2,log=FALSE)}), type="l",ylab="")
}
\keyword{CDF}
