% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPPPP_COP}
\alias{pPPPP_COP}
\title{Joint CDF of the PPPP copula model}
\usage{
pPPPP_COP(u, v, al, be, a, b)
}
\arguments{
\item{u, v:}{data input.}

\item{al, }{be, a,b:   parameters.}
}
\description{
Joint CDF of the PPPP copula model
}
\examples{
pPPPP_COP(0.9, 0.3, 1.2, 0.8, 1, 1)
}
\keyword{CDF}
\keyword{Joint}
