% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{pPPPP_COP_1_90}
\alias{pPPPP_COP_1_90}
\title{Joint CDF of the PPPP copula that is rotated for 90 degrees counter clock-wise (when a=b=1)}
\usage{
pPPPP_COP_1_90(u, v, al, be)
}
\arguments{
\item{u, v:}{data input.}

\item{al, }{be:   parameters.}
}
\description{
Joint CDF of the PPPP copula that is rotated for 90 degrees counter clock-wise (when a=b=1)
}
\examples{
pPPPP_COP_1_90(0.9, 0.3, 1.2, 0.8)
}
\keyword{CDF}
\keyword{Joint}
