% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperEV.R
\name{pUEV_GGEE_COP}
\alias{pUEV_GGEE_COP}
\title{Joint cdf of the upper extreme value copula of the GGEE copula}
\usage{
pUEV_GGEE_COP(u, v, b)
}
\arguments{
\item{u, }{v: inputs for the extreme value copula}

\item{b:}{the parameter of the extreme value copula}
}
\description{
Joint cdf of the upper extreme value copula of the GGEE copula
}
\examples{
pUEV_GGEE_COP(0.3, 0.4, 1.2)
}
\keyword{copula}
\keyword{extreme}
\keyword{upper}
\keyword{value}
