% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperEV.R
\name{pUEV_PPPP_COP}
\alias{pUEV_PPPP_COP}
\title{Joint cdf of the upper extreme value copula of the PPPP copula}
\usage{
pUEV_PPPP_COP(u, v, be, a, b)
}
\arguments{
\item{u, }{v: inputs for the extreme value copula}

\item{be, a, b:}{the parameters of the extreme value copula}
}
\description{
Joint cdf of the upper extreme value copula of the PPPP copula
}
\examples{
pUEV_PPPP_COP(0.2, 0.4, 1.2,1,1)
}
\keyword{copula}
\keyword{extreme}
\keyword{upper}
\keyword{value}
