% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{qGGEE}
\alias{qGGEE}
\title{Quantile of univariate margins of the GGEE model}
\usage{
qGGEE(u, a, b)
}
\arguments{
\item{a, b:}{parameters.}

\item{u:}{quantile between 0 and 1.}
}
\description{
Quantiel of univariate margins of the GGEE model
}
\examples{
qGGEE(0.9, 1, 1)
}
\keyword{Quantile}

