% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{qPPPP}
\alias{qPPPP}
\title{quantile of univariate margins of the PPPP model}
\usage{
qPPPP(u, al, be, a, b, maxit = 1e+05)
}
\arguments{
\item{al, be, a, b:}{parameters.}

\item{u:}{data input.}

\item{log:}{if use log in it but the result is still the original scale}
}
\description{
quantile of univariate margins of the PPPP model
}
\examples{
plot(sapply(seq(0.001,0.999, length=100), function(x){qPPPP(x, 0.3, 1.3, 1, 1,log=F)}), type="l",ylab="")
}
\keyword{quantile}
