% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{rGGEE_COP}
\alias{rGGEE_COP}
\title{Sampling based on the full-range tail dependence copula}
\usage{
rGGEE_COP(n, al, be, seed = 1)
}
\arguments{
\item{al, be:}{the two shape parameters.}

\item{n:}{sample size to be generated.}

\item{seed:}{seed for randomness, and default is 1}
}
\description{
Generating random samples based on the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
rGGEE_COP(20, 1.2, 0.2, seed = 100)
}
\keyword{simulation}
