% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{rPPPP_COP}
\alias{rPPPP_COP}
\title{Sampling based on the PPPP copula}
\usage{
rPPPP_COP(n, al, be, a, b, seed = NULL)
}
\arguments{
\item{al, }{be, a,b: the two shape parameters.}

\item{n:}{sample size to be generated.}

\item{seed:}{seed for randomness, and default is 1}
}
\description{
Generating random samples based on the PPPP copula
}
\examples{
plot(rPPPP_COP(100, 1.2, 0.2, 1, 1, seed = 100))
}
\keyword{simulation}
