% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{sprGGEE_COP}
\alias{sprGGEE_COP}
\title{Spearman's rho of the GGEE copula}
\usage{
sprGGEE_COP(al, be, flag = 1, integration = F)
}
\arguments{
\item{al, be:}{the two shape parameters.}

\item{flag:}{flag used in the appell::appellf1() function (default is -1L for auto choices)}
}
\description{
Spearman's rho of the GGEE copula
}
\examples{
sprGGEE_COP(1.2, 0.6)
}
\keyword{Spearman's}
\keyword{rho}
