% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{sprPPPP_COP}
\alias{sprPPPP_COP}
\title{Spearman's rho of the PPPP copula}
\usage{
sprPPPP_COP(al, be, a, b)
}
\arguments{
\item{al, be, a, b:}{the two shape parameters.}
}
\description{
Spearman's rho of the PPPP copula
}
\examples{
sprPPPP_COP(3, 6,5.9,6.7)
}
\keyword{Spearman's}
\keyword{rho}
