% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauGGEE_COP_sim}
\alias{tauGGEE_COP_sim}
\title{Kendall's tau of the GGEE copula (based on simulations)}
\usage{
tauGGEE_COP_sim(al, be, n = 1e+07)
}
\arguments{
\item{al, be}{the two shape parameters.}
}
\description{
Kendall's tau of the bivariate copula that has full-range tail dependence in both upper and lower tails
}
\examples{
tauGGEE_COP_sim(0.5, 1)
}
\keyword{Kendall's}
\keyword{tau}
