% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauPPPP_COP}
\alias{tauPPPP_COP}
\title{Kendall's tau of the PPPP copula}
\usage{
tauPPPP_COP(al, be, a, b)
}
\arguments{
\item{al, be, a, b}{the 4 parameters}
}
\description{
Kendall's tau of the PPPP copula
}
\examples{
tauPPPP_COP(3, 6, 5.9, 6.7)
}
\keyword{Kendall's}
\keyword{tau}
