% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{tauPPPP_COP_90}
\alias{tauPPPP_COP_90}
\title{Kendall's tau of the PPPP copula that is rotated 90 degrees counter clockwise}
\usage{
tauPPPP_COP_90(al, be, a, b, method = 1)
}
\arguments{
\item{al, be, a, b}{the 4 parameters}
}
\description{
Kendall's tau of the PPPP copula that is rotated 90 degrees counter clockwise
}
\examples{
tauPPPP_COP_90(3, 6, 5.9, 6.7)
}
\keyword{Kendall's}
\keyword{tau}
