% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullrange2.R
\name{uscore}
\alias{uscore}
\title{Uniform score function}
\usage{
uscore(data, aunif = -0.5)
}
\arguments{
\item{data:}{data to be transformed.}
}
\description{
Transform data to uniform scores, this function is from the R package CopulaModel.
}
\examples{
data  <- cbind(rnorm(10), rnorm(10))
uscore(data)
}
\keyword{score}
\keyword{uniform}

